(WIP) Split Case.tsx into a parent that just delegates to subroutes and a landing page
