+++ plugin-hrm-form/src/components/case/Case.tsx
@@ -1,5 +1,3 @@
-/* eslint-disable complexity */
-/* eslint-disable sonarjs/cognitive-complexity */
 /* eslint-disable react/prop-types */
 import React, { useEffect, useState } from 'react';
 import { Template } from '@twilio/flex-ui';
@@ -22,42 +20,26 @@ import {
 import { getConfig } from '../../HrmFormPlugin';
 import { connectToCase, transformCategories } from '../../services/ContactService';
 import { cancelCase, getActivities, updateCase } from '../../services/CaseService';
-import { completeTask, submitContactForm } from '../../services/formSubmissionHelpers';
 import { getDefinitionVersion } from '../../services/ServerlessService';
 import { getDateFromNotSavedContact, getHelplineData, isConnectedCaseActivity, sortActivities } from './caseHelpers';
-import { BottomButtonBar, Box, StyledNextStepButton } from '../../styles/HrmStyles';
-import { CaseContainer, CenteredContainer } from '../../styles/case';
-import CaseDetails from './CaseDetails';
-import { Menu, MenuItem } from '../menu';
 import { getLocaleDateTime } from '../../utils/helpers';
 import * as SearchActions from '../../states/search/actions';
 import * as CaseActions from '../../states/case/actions';
 import * as RoutingActions from '../../states/routing/actions';
 import * as ConfigActions from '../../states/configuration/actions';
-import Timeline from './Timeline';
-import CaseSummary from './CaseSummary';
 import ViewContact from './ViewContact';
-import {
-  AddTemporaryCaseInfo,
-  CaseDetailsName,
-  EditTemporaryCaseInfo,
-  TemporaryCaseInfo,
-  updateCaseListByIndex,
-  updateCaseSectionListByIndex,
-  ViewTemporaryCaseInfo,
-} from '../../states/case/types';
+import { CaseDetailsName, updateCaseListByIndex, updateCaseSectionListByIndex } from '../../states/case/types';
 import { Case as CaseType, CustomITask, NoteEntry, ReferralEntry, StandaloneITask } from '../../types/types';
 import CasePrintView from './casePrint/CasePrintView';
-import { getPermissionsForCase, PermissionActions } from '../../permissions';
-import { recordBackendError } from '../../fullStory';
-import { AppRoutes, AppRoutesWithCase, NewCaseSubroutes } from '../../states/routing/types';
-import CaseSection from './CaseSection';
-import InformationRow from './InformationRow';
-import TimelineInformationRow from './TimelineInformationRow';
-import DocumentInformationRow from './DocumentInformationRow';
+import { NewCaseSubroutes } from '../../states/routing/types';
+import CaseHome from './CaseHome';
 import AddEditCaseItem from './AddEditCaseItem';
 import ViewCaseItem from './ViewCaseItem';
 import documentUploadHandler from './documentUploadHandler';
+import { recordBackendError } from '../../fullStory';
+import { completeTask, submitContactForm } from '../../services/formSubmissionHelpers';
+import { getPermissionsForCase } from '../../permissions';
+import { CenteredContainer } from '../../styles/case';
 
 export const isStandaloneITask = (task): task is StandaloneITask => {
   return task && task.taskSid === 'standalone-task-sid';
@@ -101,17 +83,21 @@ const getFirstNameAndLastNameFromForm = (form): CaseDetailsName => {
   };
 };
 
-const splitFullName = (name: CaseDetailsName) => {
-  if (name.firstName === 'Unknown' && name.lastName === 'Unknown') {
-    return 'Unknown';
-  }
-  return `${name.firstName} ${name.lastName}`;
-};
-
-const Case: React.FC<Props> = props => {
-  const [anchorEl, setAnchorEl] = useState(null);
-  const [isMenuOpen, setMenuOpen] = useState(false);
-  const [mockedMessage, setMockedMessage] = useState(null);
+const Case: React.FC<Props> = ({
+  task,
+  form,
+  counselorsHash,
+  setConnectedCase,
+  removeConnectedCase,
+  updateCaseInfo,
+  updateCaseStatus,
+  updateCases,
+  markCaseAsUpdated,
+  changeRoute,
+  updateAllCasesView,
+  isCreating,
+  ...props
+}) => {
   const [loading, setLoading] = useState(false);
   const [timeline, setTimeline] = useState([]);
 
@@ -131,19 +117,16 @@ const Case: React.FC<Props> = props => {
       const isCreatingCase = !timelineActivities.some(isConnectedCaseActivity);
 
       if (isCreatingCase) {
-        if (isStandaloneITask(props.task)) return;
-        const date = getDateFromNotSavedContact(props.task, props.form);
+        if (isStandaloneITask(task)) return;
+        const date = getDateFromNotSavedContact(task, form);
         const { workerSid } = getConfig();
         const connectCaseActivity = {
           date: format(date, 'yyyy-MM-dd HH:mm:ss'),
           createdAt: new Date().toISOString(),
-          type: props.task.channelType,
-          text: props.form.caseInformation.callSummary.toString(),
+          type: task.channelType,
+          text: form.caseInformation.callSummary.toString(),
           twilioWorkerId: workerSid,
-          channel:
-            props.task.channelType === 'default'
-              ? props.form.contactlessTask.channel.toString()
-              : props.task.channelType,
+          channel: task.channelType === 'default' ? form.contactlessTask.channel.toString() : task.channelType,
         };
 
         timelineActivities = sortActivities([...timelineActivities, connectCaseActivity]);
@@ -152,14 +135,7 @@ const Case: React.FC<Props> = props => {
     };
 
     getTimeline();
-  }, [
-    props.task,
-    props.form,
-    props.connectedCaseId,
-    props.connectedCaseNotes,
-    props.connectedCaseReferrals,
-    setLoading,
-  ]);
+  }, [task, form, props.connectedCaseId, props.connectedCaseNotes, props.connectedCaseReferrals, setLoading]);
 
   const version = props.connectedCaseState?.connectedCase.info.definitionVersion;
   const { updateDefinitionVersion, definitionVersions } = props;
@@ -178,119 +154,25 @@ const Case: React.FC<Props> = props => {
     }
   }, [definitionVersions, updateDefinitionVersion, version]);
 
-  // Memoize can function so is not re-created on every render of Case, but only when relevant case info changes
-  const { can } = React.useMemo(
-    () =>
-      getPermissionsForCase(
-        props.connectedCaseState?.connectedCase.twilioWorkerId,
-        props.connectedCaseState?.prevStatus,
-      ),
-    [props.connectedCaseState?.connectedCase.twilioWorkerId, props.connectedCaseState?.prevStatus],
-  );
-
-  const toggleCaseMenu = e => {
-    e.persist();
-    setAnchorEl(e.currentTarget || e.target);
-    setMenuOpen(!isMenuOpen);
-  };
-
-  const handleMockedMessage = () => {
-    setMockedMessage(<Template code="NotImplemented" />);
-    setMenuOpen(false);
-  };
-
-  const closeMockedMessage = () => setMockedMessage(null);
-
-  const handleCancelNewCaseAndClose = async () => {
-    const { task } = props;
-    const { connectedCase } = props.connectedCaseState;
-    await cancelCase(connectedCase.id);
-
-    props.changeRoute({ route: 'tabbed-forms', subroute: 'caseInformation', autoFocus: true }, task.taskSid);
-    props.removeConnectedCase(task.taskSid);
-  };
-
-  const handleSaveAndEnd = async () => {
-    setLoading(true);
-
-    const { task, form } = props;
-    const { connectedCase } = props.connectedCaseState;
-    const { strings } = getConfig();
-
-    // Validating that task isn't a StandaloneITask.
-    if (isStandaloneITask(task)) return;
-
-    try {
-      const contact = await submitContactForm(task, form, connectedCase);
-      await updateCase(connectedCase.id, { ...connectedCase });
-      await connectToCase(contact.id, connectedCase.id);
-      props.markCaseAsUpdated(task.taskSid);
-      await completeTask(task);
-    } catch (error) {
-      console.error(error);
-      recordBackendError('Save and End Case', error);
-      window.alert(strings['Error-Backend']);
-    } finally {
-      setLoading(false);
-    }
-  };
-
   if (props.routing.route === 'csam-report') return null;
 
   const { route, subroute } = props.routing;
 
   // Redirects to the proper view when the user clicks 'Close' button.
   const handleClose = () => {
-    props.updateTempInfo(null, props.task.taskSid);
+    props.updateTempInfo(null, task.taskSid);
     if (route === 'select-call-type') {
-      props.changeRoute({ route: 'select-call-type' }, props.task.taskSid);
+      changeRoute({ route: 'select-call-type' }, task.taskSid);
     } else if (route === 'new-case') {
-      props.changeRoute({ route: 'new-case' }, props.task.taskSid);
+      changeRoute({ route: 'new-case' }, task.taskSid);
     } else {
-      props.changeRoute({ route: 'tabbed-forms', subroute: 'search' }, props.task.taskSid);
+      changeRoute({ route: 'tabbed-forms', subroute: 'search' }, task.taskSid);
     }
-  };
-
-  type CaseItemRoute<T extends TemporaryCaseInfo = TemporaryCaseInfo> = T['screen'] & AppRoutesWithCase['subroute'];
-
-  type CaseItemInfo<T extends TemporaryCaseInfo> = T['info']; // A bit redundant but looks cleaner than the anonymous subtype reference syntax
-
-  const onCaseItemActionClick = <T extends TemporaryCaseInfo>(
-    targetSubroute: CaseItemRoute<T>,
-  ): ((entry: CaseItemInfo<T>) => void) => (entry: CaseItemInfo<T>) => {
-    props.updateTempInfo({ screen: targetSubroute, info: entry } as TemporaryCaseInfo, props.task.taskSid);
-    props.changeRoute({ route, subroute: targetSubroute } as AppRoutes, props.task.taskSid);
-  };
-
-  const onAddCaseItemClick = (targetSubroute: CaseItemRoute<AddTemporaryCaseInfo>) => () => {
-    props.updateTempInfo({ screen: targetSubroute, info: null }, props.task.taskSid);
-    props.changeRoute({ route, subroute: targetSubroute } as AppRoutes, props.task.taskSid);
-  };
-
-  const onInfoChange = (fieldName, value) => {
-    const { connectedCase } = props.connectedCaseState;
-    const { info } = connectedCase;
-    const newInfo = info ? { ...info, [fieldName]: value } : { [fieldName]: value };
-    props.updateCaseInfo(newInfo, props.task.taskSid);
-  };
-
-  const onStatusChange = value => {
-    props.updateCaseStatus(value, props.task.taskSid);
-  };
-
-  const onClickChildIsAtRisk = () => {
-    const { connectedCase } = props.connectedCaseState;
-    const { info } = connectedCase;
-    const childIsAtRisk = !Boolean(info && info.childIsAtRisk);
-    const newInfo = info ? { ...info, childIsAtRisk } : { childIsAtRisk };
-    props.updateCaseInfo(newInfo, props.task.taskSid);
   };
 
   if (!props.connectedCaseState) return null;
 
-  const { task, form, counselorsHash } = props;
-
-  const { connectedCase, caseHasBeenEdited, prevStatus } = props.connectedCaseState;
+  const { connectedCase, prevStatus, caseHasBeenEdited } = props.connectedCaseState;
 
   const getCategories = firstConnectedContact => {
     if (firstConnectedContact?.rawJson?.caseInformation) {
@@ -302,42 +184,16 @@ const Case: React.FC<Props> = props => {
     return null;
   };
 
-  const handleUpdate = async () => {
-    setLoading(true);
-    const { strings } = getConfig();
-
-    try {
-      const updatedCase = await updateCase(connectedCase.id, { ...connectedCase });
-      props.setConnectedCase(updatedCase, task.taskSid, false);
-      props.updateCases(task.taskSid, updatedCase);
-      // IF case has been edited from All Cases view, we should update that view
-      if (props.updateAllCasesView) {
-        props.updateAllCasesView(updatedCase);
-      }
-    } catch (error) {
-      console.error(error);
-      recordBackendError('Update Case', error);
-      window.alert(strings['Error-Backend']);
-    } finally {
-      setLoading(false);
-    }
-  };
-
-  const onPrintCase = () => {
-    props.changeRoute({ route, subroute: 'case-print-view' }, task.taskSid);
-  };
-
-  const isMockedMessageOpen = Boolean(mockedMessage);
+  const { can } = getPermissionsForCase(connectedCase.twilioWorkerId, prevStatus);
 
   const firstConnectedContact = connectedCase && connectedCase.connectedContacts && connectedCase.connectedContacts[0];
   const name = firstConnectedContact
     ? getFirstNameAndLastNameFromContact(firstConnectedContact)
     : getFirstNameAndLastNameFromForm(form);
 
-  const fullName = splitFullName(name);
   const categories = getCategories(firstConnectedContact);
   const { createdAt, updatedAt, twilioWorkerId, status, info } = connectedCase || {};
-  const { workerSid, featureFlags } = getConfig();
+  const { workerSid } = getConfig();
   const caseCounselor = counselorsHash[twilioWorkerId];
   const currentCounselor = counselorsHash[workerSid];
   const openedDate = getLocaleDateTime(createdAt);
@@ -357,8 +213,68 @@ const Case: React.FC<Props> = props => {
   const definitionVersion = props.definitionVersions[version];
   const office = getHelplineData(connectedCase.helpline, definitionVersion.helplineInformation);
 
+  const onInfoChange = (fieldName, value) => {
+    const newInfo = info ? { ...info, [fieldName]: value } : { [fieldName]: value };
+    updateCaseInfo(newInfo, task.taskSid);
+  };
+
+  const onStatusChange = (value: string) => {
+    updateCaseStatus(value, task.taskSid);
+  };
+
+  const handleUpdate = async () => {
+    setLoading(true);
+    const { strings } = getConfig();
+
+    try {
+      const updatedCase = await updateCase(connectedCase.id, { ...connectedCase });
+      setConnectedCase(updatedCase, task.taskSid, false);
+      updateCases(task.taskSid, updatedCase);
+      // IF case has been edited from All Cases view, we should update that view
+      if (updateAllCasesView) {
+        updateAllCasesView(updatedCase);
+      }
+    } catch (error) {
+      console.error(error);
+      recordBackendError('Update Case', error);
+      window.alert(strings['Error-Backend']);
+    } finally {
+      setLoading(false);
+    }
+  };
+
+  const handleCancelNewCaseAndClose = async () => {
+    await cancelCase(connectedCase.id);
+
+    changeRoute({ route: 'tabbed-forms', subroute: 'caseInformation', autoFocus: true }, task.taskSid);
+    removeConnectedCase(task.taskSid);
+  };
+
+  const handleSaveAndEnd = async () => {
+    setLoading(true);
+
+    const { strings } = getConfig();
+
+    // Validating that task isn't a StandaloneITask.
+    if (isStandaloneITask(task)) return;
+
+    try {
+      const contact = await submitContactForm(task, form, connectedCase);
+      await updateCase(connectedCase.id, { ...connectedCase });
+      await connectToCase(contact.id, connectedCase.id);
+      markCaseAsUpdated(task.taskSid);
+      await completeTask(task);
+    } catch (error) {
+      console.error(error);
+      recordBackendError('Save and End Case', error);
+      window.alert(strings['Error-Backend']);
+    } finally {
+      setLoading(false);
+    }
+  };
+
   const addScreenProps = {
-    task: props.task,
+    task,
     route: props.routing.route,
     counselor: currentCounselor,
     counselorsHash,
@@ -374,11 +290,13 @@ const Case: React.FC<Props> = props => {
     name,
     categories,
     status,
+    prevStatus,
     caseCounselor,
     currentCounselor,
     openedDate,
     lastUpdatedDate,
-    followUpDate: followUpPrintedDate,
+    followUpDate,
+    followUpPrintedDate,
     households,
     perpetrators,
     incidents,
@@ -390,87 +308,9 @@ const Case: React.FC<Props> = props => {
     office,
     version,
     contact: firstConnectedContact,
+    contacts: connectedCase?.connectedContacts ?? [],
   };
 
-  const itemRowRenderer = <TViewCaseInfo extends TemporaryCaseInfo>(
-    itemTypeName: string,
-    viewSubroute: CaseItemRoute,
-    items: CaseItemInfo<TViewCaseInfo>[],
-  ) => {
-    const itemRows = () => {
-      return items.length ? (
-        <>
-          {items.map((item, index) => (
-            <InformationRow
-              key={`${itemTypeName}-${index}`}
-              person={item[itemTypeName]}
-              onClickView={() => onCaseItemActionClick<TViewCaseInfo>(viewSubroute)(item)}
-            />
-          ))}
-        </>
-      ) : null;
-    };
-    itemRows.displayName = `${name}Rows`;
-    return itemRows;
-  };
-
-  const householdRows = itemRowRenderer<ViewTemporaryCaseInfo>(
-    'form',
-    NewCaseSubroutes.ViewHousehold,
-    households.map((h, index) => {
-      const { household, ...caseInfoItem } = { ...h, form: h.household, id: null };
-      return { ...caseInfoItem, index };
-    }),
-  );
-
-  const perpetratorRows = itemRowRenderer<ViewTemporaryCaseInfo>(
-    'form',
-    NewCaseSubroutes.ViewPerpetrator,
-    perpetrators.map((p, index) => {
-      const { perpetrator, ...caseInfoItem } = { ...p, form: p.perpetrator, id: null };
-      return { ...caseInfoItem, index };
-    }),
-  );
-
-  const incidentRows = () => {
-    return incidents.length ? (
-      <>
-        {incidents.map((item, index) => {
-          const { incident, ...caseItemEntry } = { ...item, form: item.incident, id: null };
-          return (
-            <TimelineInformationRow
-              key={`incident-${index}`}
-              onClickView={() =>
-                onCaseItemActionClick<ViewTemporaryCaseInfo>(NewCaseSubroutes.ViewIncident)({ ...caseItemEntry, index })
-              }
-              definition={caseForms.IncidentForm}
-              values={item.incident}
-              layoutDefinition={caseLayouts.incidents}
-            />
-          );
-        })}
-      </>
-    ) : null;
-  };
-
-  const documentRows = () => {
-    return documents.length ? (
-      <>
-        {documents.map((item, index) => {
-          const { document, ...caseItemEntry } = { ...item, form: item.document, index };
-          return (
-            <DocumentInformationRow
-              key={`document-${index}`}
-              documentEntry={item}
-              onClickView={() =>
-                onCaseItemActionClick<ViewTemporaryCaseInfo>(NewCaseSubroutes.ViewDocument)(caseItemEntry)
-              }
-            />
-          );
-        })}
-      </>
-    ) : null;
-  };
   switch (subroute) {
     case NewCaseSubroutes.AddNote:
     case NewCaseSubroutes.EditNote:
@@ -605,126 +445,21 @@ const Case: React.FC<Props> = props => {
           <CircularProgress size={50} />
         </CenteredContainer>
       ) : (
-        <>
-          <CaseContainer>
-            <Box marginLeft="25px" marginTop="13px">
-              <CaseDetails
-                caseId={connectedCase.id}
-                name={fullName}
-                status={status}
-                prevStatus={prevStatus}
-                can={can}
-                counselor={caseCounselor}
-                categories={categories}
-                openedDate={openedDate}
-                lastUpdatedDate={lastUpdatedDate}
-                followUpDate={followUpDate}
-                childIsAtRisk={childIsAtRisk}
-                office={office?.label}
-                handlePrintCase={onPrintCase}
-                handleInfoChange={onInfoChange}
-                handleStatusChange={onStatusChange}
-                handleClickChildIsAtRisk={onClickChildIsAtRisk}
-                definitionVersion={definitionVersion}
-                definitionVersionName={connectedCase.info.definitionVersion}
-                isOrphanedCase={!firstConnectedContact}
-              />
-            </Box>
-            <Box marginLeft="25px" marginTop="25px">
-              <Timeline
-                timelineActivities={timeline}
-                caseObj={connectedCase}
-                taskSid={task.taskSid}
-                form={form}
-                can={can}
-                route={route}
-              />
-            </Box>
-            <Box marginLeft="25px" marginTop="25px">
-              <CaseSection
-                can={() => can(PermissionActions.ADD_HOUSEHOLD)}
-                onClickAddItem={onAddCaseItemClick(NewCaseSubroutes.AddHousehold)}
-                sectionTypeId="Household"
-              >
-                {householdRows()}
-              </CaseSection>
-            </Box>
-            <Box marginLeft="25px" marginTop="25px">
-              <CaseSection
-                can={() => can(PermissionActions.ADD_PERPETRATOR)}
-                onClickAddItem={onAddCaseItemClick(NewCaseSubroutes.AddPerpetrator)}
-                sectionTypeId="Perpetrator"
-              >
-                {perpetratorRows()}
-              </CaseSection>
-            </Box>
-            <Box marginLeft="25px" marginTop="25px">
-              <CaseSection
-                can={() => can(PermissionActions.ADD_INCIDENT)}
-                onClickAddItem={onAddCaseItemClick(NewCaseSubroutes.AddIncident)}
-                sectionTypeId="Incident"
-              >
-                {incidentRows()}
-              </CaseSection>
-            </Box>
-            {featureFlags.enable_upload_documents && (
-              <Box marginLeft="25px" marginTop="25px">
-                <CaseSection
-                  onClickAddItem={onAddCaseItemClick(NewCaseSubroutes.AddDocument)}
-                  can={() => can(PermissionActions.ADD_DOCUMENT)}
-                  sectionTypeId="Document"
-                >
-                  {documentRows()}
-                </CaseSection>
-              </Box>
-            )}
-            <Box marginLeft="25px" marginTop="25px">
-              <CaseSummary task={props.task} readonly={!can(PermissionActions.EDIT_CASE_SUMMARY)} />
-            </Box>
-            <Dialog onClose={closeMockedMessage} open={isMockedMessageOpen}>
-              <DialogContent>{mockedMessage}</DialogContent>
-            </Dialog>
-            <Menu anchorEl={anchorEl} open={isMenuOpen} onClickAway={() => setMenuOpen(false)}>
-              <MenuItem
-                Icon={AddIcon}
-                text={<Template code="BottomBar-AddThisContactToExistingCase" />}
-                onClick={handleMockedMessage}
-              />
-              <MenuItem
-                red
-                Icon={CancelIcon}
-                text={<Template code="BottomBar-CancelNewCaseAndClose" />}
-                onClick={handleCancelNewCaseAndClose}
-              />
-            </Menu>
-          </CaseContainer>
-          <BottomButtonBar>
-            {props.isCreating && (
-              <>
-                <Box marginRight="15px">
-                  <StyledNextStepButton secondary roundCorners onClick={toggleCaseMenu}>
-                    <Template code="BottomBar-Cancel" />
-                  </StyledNextStepButton>
-                </Box>
-                <StyledNextStepButton roundCorners onClick={handleSaveAndEnd}>
-                  <Template code="BottomBar-SaveAndEnd" />
-                </StyledNextStepButton>
-              </>
-            )}
-            {!props.isCreating && (
-              <>
-                <Box marginRight="15px">
-                  <StyledNextStepButton secondary roundCorners onClick={props.handleClose}>
-                    <Template code="BottomBar-Close" />
-                  </StyledNextStepButton>
-                </Box>
-                <StyledNextStepButton disabled={!caseHasBeenEdited} roundCorners onClick={handleUpdate}>
-                  <Template code="BottomBar-Update" />
-                </StyledNextStepButton>
-              </>
-            )}
-          </BottomButtonBar>
-        </>
+        <CaseHome
+          task={task}
+          definitionVersion={definitionVersion}
+          caseDetails={caseDetails}
+          timeline={timeline}
+          handleClose={handleClose}
+          handleCancelNewCaseAndClose={handleCancelNewCaseAndClose}
+          handleUpdate={handleUpdate}
+          handleSaveAndEnd={handleSaveAndEnd}
+          onInfoChange={onInfoChange}
+          onStatusChange={onStatusChange}
+          isCreating={isCreating}
+          isEdited={Boolean(caseHasBeenEdited)}
+          can={can}
+        />
       );
   }
 };
