+++ plugin-hrm-form/src/components/case/casePrint/CasePrintView.tsx
@@ -130,7 +130,7 @@ const CasePrintView: React.FC<Props> = ({ onClickClose, caseDetails, definitionV
                   status={caseDetails.status}
                   openedDate={caseDetails.openedDate}
                   lastUpdatedDate={caseDetails.lastUpdatedDate}
-                  followUpDate={caseDetails.followUpDate}
+                  followUpDate={caseDetails.followUpPrintedDate}
                   childIsAtRisk={caseDetails.childIsAtRisk}
                   counselor={caseDetails.caseCounselor}
                   categories={caseDetails.categories}
